Combined the two math translation classes to a single one
